Factored out Positioned into separate file.
